Prevent treadmilling in CopyDirectory (#10)
Internal testing showed that
https://github.com/Qumulo/qumulo-filesystem-walk/pull/8 made it possible
or at least much more likely for CopyDirectory to get stuck in a loop
waiting for files to be created under a directory before the directory
attributes were applied when MAX_QUEUE_LEN is small.

This restores the old behavior but only for CopyDirectory. The queue
elements for CopyDirectory are just `int`'s, so very small.